Implement add and remove
