--- src/modules/__mocks__/elements.js
@@ -1,14 +1,0 @@
-const descInput = {
-  value: 'sometext',
-};
-const clearCompleted = document.querySelector('.btn-clear');
-const statusInput = document.getElementsByClassName('status');
-const todoList = document.querySelector('.todo-lists-holder');
-const editButton = document.getElementsByClassName('btn-edit');
-const deleteButton = document.getElementsByClassName('btn-delete');
-const form = document.getElementsByTagName('form')[0];
-const refreshButton = document.querySelector('.btn-refresh');
-
-export {
-  descInput, clearCompleted, statusInput, todoList, editButton, deleteButton, form, refreshButton,
-};
